Split lisp code
